requirements.txt
numpy==1.26.4
opencv-python==4.11.0.86
mediapipe==0.10.21
scikit-learn>=1.3.0
openpyxl>=3.1.0
Pillow>=10.0.0
vosk>=0.3.45
pyserial>=3.5
anthropic>=0.25.0
urllib3>=2.0.0
